+++ src/main/java/org/tptacs/presentation/dto/ItemDto.java
@@ -1,5 +1,7 @@
 package org.tptacs.presentation.dto;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -9,10 +11,10 @@ public class ItemDto {
 	@JsonProperty("name")
     private String name;
 	@JsonProperty("price")
-    private Double price;
+    private BigDecimal price;
 	
 	@JsonCreator
-	public ItemDto(String id, String name, Double price) {
+	public ItemDto(String id, String name, BigDecimal price) {
 		this.id = id;
 		this.name = name;
 		this.price = price;
